requirements.txt
telethon==1.35.0
pandas==2.2.2
python-dotenv==1.0.1
datasets==2.19.1
transformers==4.41.0
scikit-learn==1.5.0
matplotlib==3.9.0
seaborn==0.13.2
jupyter==1.0.0
